Flask_foled_structure/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
passlib
flask-bcrypt
email-validator